+++ src/compas_shapeop/solvers/__init__.py
@@ -1,0 +1,3 @@
+from .meshsolver import MeshSolver
+
+__all__ = ["MeshSolver"]