+++ tests/test_dynamicsolver.py
@@ -9,12 +9,15 @@ def test_solver_basic_functionality():
     rows, cols = 14, 14
     mesh = Mesh.from_meshgrid(nx=cols - 1, ny=rows - 1, dx=1.0 * (cols - 1), dy=1.0 * (rows - 1))
     mesh.translate([-5, -5, 0])
+    points = mesh.to_vertices_and_faces()[0]
 
-    solver = Solver.from_mesh(mesh)
+    solver = Solver()
+    solver.points = points
     corner_vertices = [v for v in mesh.vertices() if len(mesh.vertex_neighbors(v)) == 2]
     solver.add_closeness_constraints(corner_vertices, weight=1e5)
-    solver.add_mesh_edge_strain_constraint(mesh, weight=1.0, min_range=0.8, max_range=1.1)
-    solver.add_mesh_vertex_force(mesh, force_x=0, force_y=0, force_z=0.001)
+    for edge in mesh.edges():
+        solver.add_edge_strain_constraint(edge[0], edge[1], weight=1.0, min_range=0.8, max_range=1.1)
+    solver.add_gravity_force(0, 0, 0.001)
     solver.init()
     original_positions = solver.points.copy()
     solver.solve(20)